+++ packages/cubejs-client-vue3/src/index.js
@@ -1,0 +1,8 @@
+import { GRANULARITIES } from '@cubejs-client/core';
+
+import QueryRenderer from './QueryRenderer';
+import QueryBuilder from './QueryBuilder';
+
+export { QueryRenderer, QueryBuilder, GRANULARITIES };
+
+export default {};