+++ rust/cubestore-sql-tests/src/tests.rs
@@ -2,8 +2,10 @@ use crate::SqlClient;
 use async_compression::tokio::write::GzipEncoder;
 use cubestore::queryplanner::pretty_printers::{pp_phys_plan, pp_phys_plan_ext, PPOptions};
 use cubestore::queryplanner::MIN_TOPK_STREAM_ROWS;
+use cubestore::sql::timestamp_from_string;
 use cubestore::store::DataFrame;
 use cubestore::table::{Row, TableValue, TimestampValue};
+use cubestore::util::decimal::Decimal;
 use itertools::Itertools;
 use pretty_assertions::assert_eq;
 use std::env;
@@ -73,6 +75,9 @@ pub fn sql_tests() -> Vec<(&'static str, TestFn)> {
         t("topk_decimals", topk_decimals),
         t("offset", offset),
         t("having", having),
+        t("rolling_window_join", rolling_window_join),
+        t("decimal_index", decimal_index),
+        t("float_index", float_index),
     ];
 
     fn t<F>(name: &'static str, f: fn(Box<dyn SqlClient>) -> F) -> (&'static str, TestFn)
@@ -283,11 +288,11 @@ async fn group_by_decimal(service: Box<dyn SqlClient>) {
         result.get_rows(),
         &vec![
             Row::new(vec![
-                TableValue::Decimal("100".to_string()),
+                TableValue::Decimal(Decimal::new(100 * 100_000)),
                 TableValue::Int(3)
             ]),
             Row::new(vec![
-                TableValue::Decimal("200".to_string()),
+                TableValue::Decimal(Decimal::new(200 * 100_000)),
                 TableValue::Int(2)
             ])
         ]
@@ -2106,7 +2111,7 @@ async fn topk_decimals(service: Box<dyn SqlClient>) {
             .map(|(s, i)| {
                 vec![
                     TableValue::String(s.to_string()),
-                    TableValue::Decimal(i.to_string()),
+                    TableValue::Decimal(Decimal::new(*i * 100_000)),
                 ]
             })
             .collect_vec()
@@ -2212,22 +2217,229 @@ async fn having(service: Box<dyn SqlClient>) {
         .unwrap();
     assert_eq!(to_rows(&r), rows(&[("a", 2), ("b", 1)]));
 
+    // We diverge from datafusion here, which resolve `n` in the HAVING to `sum(n)` and fail.
+    // At the moment CubeJS sends requests like this, though, so we choose to remove support for
+    // filtering on aliases in the same query.
     let r = service
         .exec_query(
-            "SELECT `data`.id, count(`data`.n) `cnt` \
+            "SELECT `data`.id, sum(n) AS n \
              FROM (SELECT * FROM s.Data1 UNION ALL SELECT * FROM s.Data2) `data` \
-             WHERE n != 2 \
              GROUP BY 1 \
-             HAVING cnt = 2 \
+             HAVING sum(n) > 5 \
              ORDER BY 1",
         )
         .await
         .unwrap();
-    assert_eq!(to_rows(&r), rows(&[("a", 2), ("c", 2)]));
+    assert_eq!(to_rows(&r), rows(&[("b", 7), ("c", 9)]));
+    // Since we do not resolve aliases, this will fail.
+    let err = service
+        .exec_query(
+            "SELECT `data`.id, sum(n) AS n \
+             FROM (SELECT * FROM s.Data1 UNION ALL SELECT * FROM s.Data2) `data` \
+             GROUP BY 1 \
+             HAVING n = 2 \
+             ORDER BY 1",
+        )
+        .await;
+    assert!(err.is_err());
 
     fn rows(a: &[(&str, i64)]) -> Vec<Vec<TableValue>> {
         a.iter()
             .map(|(s, n)| vec![TableValue::String(s.to_string()), TableValue::Int(*n)])
+            .collect_vec()
+    }
+}
+
+async fn rolling_window_join(service: Box<dyn SqlClient>) {
+    service.exec_query("CREATE SCHEMA s").await.unwrap();
+    service
+        .exec_query("CREATE TABLE s.Data(day timestamp, name text, n int)")
+        .await
+        .unwrap();
+    let raw_query = "SELECT Series.date_to, Table.name, sum(Table.n) as n FROM (\
+               SELECT to_timestamp('2020-01-01T00:00:00.000') date_from, \
+                      to_timestamp('2020-01-01T23:59:59.999') date_to \
+               UNION ALL \
+               SELECT to_timestamp('2020-01-02T00:00:00.000') date_from, \
+                      to_timestamp('2020-01-02T23:59:59.999') date_to \
+               UNION ALL \
+               SELECT to_timestamp('2020-01-03T00:00:00.000') date_from, \
+                      to_timestamp('2020-01-03T23:59:59.999') date_to \
+               UNION ALL \
+               SELECT to_timestamp('2020-01-04T00:00:00.000') date_from, \
+                      to_timestamp('2020-01-04T23:59:59.999') date_to\
+            ) AS `Series` \
+            LEFT JOIN (\
+               SELECT date_trunc('day', CONVERT_TZ(day,'+00:00')) `day`, name, sum(n) `n` \
+               FROM s.Data \
+               GROUP BY 1, 2 \
+            ) AS `Table` ON `Table`.day <= `Series`.date_to \
+            GROUP BY 1, 2";
+    let query = raw_query.to_string() + " ORDER BY 1, 2, 3";
+    let query_sort_subquery = format!(
+        "SELECT q0.date_to, q0.name, q0.n FROM ({}) as q0 ORDER BY 1,2,3",
+        raw_query
+    );
+
+    let plan = service.plan_query(&query).await.unwrap().worker;
+    assert_eq!(
+        pp_phys_plan(plan.as_ref()),
+        "Sort\
+      \n  Projection, [date_to, name, SUM(n):n]\
+      \n    CrossJoinAgg, on: day <= date_to\
+      \n      Alias\
+      \n        Projection, [day, name, SUM(n):n]\
+      \n          FinalHashAggregate\
+      \n            Worker\
+      \n              PartialHashAggregate\
+      \n                Merge\
+      \n                  Scan, index: default:1:[1], fields: *\
+      \n                    Empty"
+    );
+
+    let plan = service
+        .plan_query(&query_sort_subquery)
+        .await
+        .unwrap()
+        .worker;
+    assert_eq!(
+        pp_phys_plan(plan.as_ref()),
+        "Sort\
+        \n  Projection, [date_to, name, n]\
+        \n    Alias\
+        \n      Projection, [date_to, name, SUM(n):n]\
+        \n        CrossJoinAgg, on: day <= date_to\
+        \n          Alias\
+        \n            Projection, [day, name, SUM(n):n]\
+        \n              FinalHashAggregate\
+        \n                Worker\
+        \n                  PartialHashAggregate\
+        \n                    Merge\
+        \n                      Scan, index: default:1:[1], fields: *\
+        \n                        Empty"
+    );
+
+    service
+        .exec_query("INSERT INTO s.Data(day, name, n) VALUES ('2020-01-01T01:00:00.000', 'john', 10), \
+                                                             ('2020-01-01T01:00:00.000', 'sara', 7), \
+                                                             ('2020-01-03T02:00:00.000', 'sara', 3), \
+                                                             ('2020-01-03T03:00:00.000', 'john', 9), \
+                                                             ('2020-01-03T03:00:00.000', 'john', 11), \
+                                                             ('2020-01-04T05:00:00.000', 'timmy', 5)")
+        .await
+        .unwrap();
+
+    let mut jan = (1..=4)
+        .map(|d| timestamp_from_string(&format!("2020-01-{:02}T23:59:59.999", d)).unwrap())
+        .collect_vec();
+    jan.insert(0, jan[1]); // jan[i] will correspond to i-th day of the month.
+
+    for q in &[query.as_str(), query_sort_subquery.as_str()] {
+        log::info!("Testing query {}", q);
+        let r = service.exec_query(q).await.unwrap();
+        assert_eq!(
+            to_rows(&r),
+            rows(&[
+                (jan[1], "john", 10),
+                (jan[1], "sara", 7),
+                (jan[2], "john", 10),
+                (jan[2], "sara", 7),
+                (jan[3], "john", 30),
+                (jan[3], "sara", 10),
+                (jan[4], "john", 30),
+                (jan[4], "sara", 10),
+                (jan[4], "timmy", 5)
+            ])
+        );
+    }
+
+    fn rows(a: &[(TimestampValue, &str, i64)]) -> Vec<Vec<TableValue>> {
+        a.iter()
+            .map(|(t, s, n)| {
+                vec![
+                    TableValue::Timestamp(*t),
+                    TableValue::String(s.to_string()),
+                    TableValue::Int(*n),
+                ]
+            })
+            .collect_vec()
+    }
+}
+
+async fn decimal_index(service: Box<dyn SqlClient>) {
+    service.exec_query("CREATE SCHEMA s").await.unwrap();
+    service
+        .exec_query("CREATE TABLE s.Data(x decimal, y decimal)")
+        .await
+        .unwrap();
+    service
+        .exec_query("CREATE INDEX reverse on s.Data(y, x)")
+        .await
+        .unwrap();
+    service
+        .exec_query("INSERT INTO s.Data(x,y) VALUES (1, 2), (2, 3), (3, 4)")
+        .await
+        .unwrap();
+
+    let r = service
+        .exec_query("SELECT * FROM s.Data ORDER BY x")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&[(1, 2), (2, 3), (3, 4)]));
+
+    let r = service
+        .exec_query("SELECT * FROM s.Data ORDER BY y DESC")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&[(3, 4), (2, 3), (1, 2)]));
+
+    fn rows(a: &[(i64, i64)]) -> Vec<Vec<TableValue>> {
+        a.iter()
+            .map(|(x, y)| {
+                vec![
+                    TableValue::Decimal(Decimal::new(x * 100000)),
+                    TableValue::Decimal(Decimal::new(y * 100000)),
+                ]
+            })
+            .collect_vec()
+    }
+}
+
+async fn float_index(service: Box<dyn SqlClient>) {
+    service.exec_query("CREATE SCHEMA s").await.unwrap();
+    service
+        .exec_query("CREATE TABLE s.Data(x float, y float)")
+        .await
+        .unwrap();
+    service
+        .exec_query("CREATE INDEX reverse on s.Data(y, x)")
+        .await
+        .unwrap();
+    service
+        .exec_query("INSERT INTO s.Data(x,y) VALUES (1, 2), (2, 3), (3, 4)")
+        .await
+        .unwrap();
+
+    let r = service
+        .exec_query("SELECT * FROM s.Data ORDER BY x")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&[(1., 2.), (2., 3.), (3., 4.)]));
+
+    let r = service
+        .exec_query("SELECT * FROM s.Data ORDER BY y DESC")
+        .await
+        .unwrap();
+    assert_eq!(to_rows(&r), rows(&[(3., 4.), (2., 3.), (1., 2.)]));
+
+    fn rows(a: &[(f64, f64)]) -> Vec<Vec<TableValue>> {
+        a.iter()
+            .map(|(x, y)| {
+                vec![
+                    TableValue::Float((*x).into()),
+                    TableValue::Float((*y).into()),
+                ]
+            })
             .collect_vec()
     }
 }