+++ rust/cubestore/src/queryplanner/optimizations/rewrite_plan.rs
@@ -109,7 +109,16 @@ pub fn rewrite_plan<R: PlanRewriter>(
             stringified_plans: stringified_plans.clone(),
             schema: schema.clone(),
         },
-        p @ LogicalPlan::Extension { .. } => p.clone(),
+        LogicalPlan::Extension { node } => LogicalPlan::Extension {
+            node: node.from_template(
+                &node.expressions(),
+                &node
+                    .inputs()
+                    .into_iter()
+                    .map(|p| rewrite_plan(p, ctx, f))
+                    .collect::<Result<Vec<_>, _>>()?,
+            ),
+        },
     };
 
     // Update the resulting plan.